requirements.txt
scrapy #==2.11.0
selenium  #==4.15.0
pandas #==2.2.2
plotly #==5.18.0
dash #==2.14.0
psycopg2-binary #==2.9.7
sqlalchemy #==2.0.23
requests #==2.31.0
scrapy-rotating-proxies #==0.3.10
python-dotenv #==1.0.0
webdriver-manager #==4.0.1
tenacity #==8.2.2
beautifulsoup4 #==4.12.2